+++ src/kannushi/_cli/__init__.py
@@ -1,0 +1,7 @@
+__all__ = ["main"]
+
+#
+# Package-level exports
+#
+
+from .main import main